requirements.txt
uvicorn>=0.16.0,<0.18.0
fastapi>=0.64.0,<1.0.0
pydantic #>=1.8.1,<2.0.0
requests>=2.23.0,<2.24.0

# If locally copied whl file inside bikeshare_model_api then use given below
bikeshare_model-0.0.1-py3-none-any.whl

# After uploding your package in PyPI use the given below
#bikeshare-model==0.0.1